backend/risk_map/utils/face_alignment.py
"""
Face alignment utilities for Medical AI Assistant
"""
import logging
import numpy as np
from PIL import Image

logger = logging.getLogger(__name__)

class FaceAligner:
    """Simple face alignment utility for standardized processing"""
    
    def __init__(self):
        self.logger = logging.getLogger(self.__class__.__name__)
        
    def normalize_face(self, image, landmarks=None):
        """
        Normalize face for consistent processing
        
        Args:
            image: PIL Image or numpy array
            landmarks: Optional landmark points
            
        Returns:
            Normalized image (for now, just return original)
        """
        try:
            # Simple stub implementation - just return the image
            # In production, this would do proper face alignment
            
            if isinstance(image, np.ndarray):
                # Convert to PIL if needed
                if len(image.shape) == 3:
                    image = Image.fromarray(image)
                elif len(image.shape) == 2:
                    image = Image.fromarray(image, mode='L')
            
            self.logger.info("Face normalization completed (stub implementation)")
            return image
            
        except Exception as e:
            self.logger.error(f"Face normalization failed: {e}")
            return image  # Return original on error