requirements-cloudrun.txt
# Cloud Run Requirements for NuvaFace API
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
pydantic>=2.0.0
python-multipart>=0.0.5
Pillow>=9.0.0
numpy>=1.21.0
opencv-python-headless>=4.5.0
mediapipe>=0.8.0
python-dotenv>=0.19.0
google-genai>=0.7.0

# Simplified dependencies - remove heavy ML libs for faster build
# torch>=1.9.0
# torchvision>=0.10.0
# insightface>=0.2.1
# onnxruntime>=1.8.0